+++ app/build.gradle
@@ -1,6 +1,9 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id("androidx.navigation.safeargs.kotlin")
+    id 'kotlin-kapt'
+
 }
 
 android {
@@ -45,8 +48,15 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     implementation "com.mikepenz:materialdrawer:7.0.0"
+    implementation "com.mikepenz:materialdrawer-nav:7.0.0"
+
     implementation "androidx.recyclerview:recyclerview:1.2.1"
     implementation "androidx.annotation:annotation:1.5.0"
 
     implementation 'de.hdodenhof:circleimageview:3.1.0'
+
+    implementation 'com.fragula2:fragula-core:2.4.1'
+
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
 }
